refactor(send): isolate received hook planning
Findings: RULE-002-HTTPPERF, RULE-003-HTTPPERF
